Separate helpers from api
